requirements.txt
opencv-python
tqdm
torch
transformers
sentence-transformers
numpy
google-generativeai   # optional
